+++ Tests/KeyboardKitTests/Context/ObservableKeyboardContextTests.swift
@@ -30,7 +30,6 @@ class ObservableKeyboardContextTests: QuickSpec {
                 let context = ObservableKeyboardContext(from: standard)
                 expect(context.actionHandler).to(be(standard.actionHandler))
                 expect(context.controller).to(be(standard.controller))
-                expect(context.emojiCategory).to(equal(.animals))
                 expect(context.hasDictationKey).to(equal(standard.hasDictationKey))
                 expect(context.hasFullAccess).to(equal(standard.hasFullAccess))
                 expect(context.keyboardInputSetProvider).to(be(standard.keyboardInputSetProvider))